Introduce CampaignService
