+++ src/application/campaign_service.go
@@ -1,0 +1,14 @@
+package application
+
+type IEmailRepository interface {
+	AddEmail(email string) error
+	GetAll() []string
+	Save() error
+}
+
+type IEmailClient interface {
+	Send(recipients []string, htmlContent string) error
+}
+
+type CampaignService struct {
+}
